--- src/main/java/com/capgemini/ccsw/ccswmanager/Usuario/UsuarioService.java
@@ -1,10 +1,0 @@
-package com.capgemini.ccsw.ccswmanager.usuario;
-
-import java.util.List;
-
-import com.capgemini.ccsw.ccswmanager.usuario.model.Usuario;
-
-public interface UsuarioService {
-  
-  List<Usuario> findUsuario();
-}
