+++ src/main/java/com/capgemini/ccsw/ccswmanager/usuario2/UsuarioService.java
@@ -1,0 +1,10 @@
+package com.capgemini.ccsw.ccswmanager.usuario;
+
+import java.util.List;
+
+import com.capgemini.ccsw.ccswmanager.usuario.model.Usuario;
+
+public interface UsuarioService {
+  
+  List<Usuario> findUsuario();
+}
